Fix export when no location has flag prestashop_synchronized
When there is no location with the flag prestashop_synchronized in the
location tree,

Current behavior:

The `location` key in the context is empty then the qty_available is
computed for *every* locations/warehouses.

Expected behavior:

If there is at least one location flagged, use it, otherwise, use all
the locations of the tree, starting from the selected location or
selected warehouse location.

When no location can be found (for instance because user selected a
location with a usage different than 'internal'), then it should just
fail and not return the stock of all warehouses.

Being able to synchronize without activating 'prestashop_synchronized'
is important: when used, this feature will trigger an export every time
a quant is changed or created, on a large catalog, we might prefer only
using the cron to have less frequent updates and less jobs.

A second fix is that when a stock_location_id is set on the backend, it
should be used, it was only used for the import of stock.